renames project to rime
