+++ src/rime/sta/__init__.py
