+++ src/rime/sta/core.py
