+++ src/rime/sta/schema.py
