+++ tests/rime/frost/test_post.py
@@ -15,7 +15,7 @@ import requests
 
 from rime.frost.post import general_post, make_frost_entity
 from rime.frost.types import FrostEntityRef
-from rime.sensor_things.core import Thing
+from rime.sta.core import Thing
 
 if TYPE_CHECKING:
     from conftest import DockerFrost
