+++ tests/rime/sta/data/empty_iot_link_sensor_config.yaml
