+++ tests/rime/sta/data/valid_sensor_config.yaml
